requirements.txt
Flask
requests
python-dotenv
Flask-Spotify-Auth